pyproject.toml
[tool.poetry]
name = "challenger"
version = "0.1.0"
description = ""
authors = ["konstantin <kmerezhnikov2002@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
djangorestframework = "^3.14.0"
python-dotenv = "^0.21.0"
requests = "^2.28.1"
websockets = "^10.4"
dj-database-url = "^1.0.0"
psycopg2-binary = "^2.9.5"
pillow = "^9.3.0"
drf-yasg = "^1.21.4"
django-filter = "^22.1"
whitenoise = "^6.2.0"
gunicorn = "^20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
